Reworked opearators:
Added 3 new functions in the 'internal' namespace that replace all usages of standard comparison
operators. These new functions behave the same as standard operators for every type
that is not a floating type. For floating types, comparison is done such that two
numbers are treated as equal if the absolute difference between them is no greater than a set epsilon value