Fixed deleteEdge and addEdge member functions
In case of undirected graphs, deleteEdge would only delete one direction of the edge, not both.
In case of adding and edge (with any of the addEdge overloads) where the starting node is already
in the adjacency list, and end node does not exist, it would only be added in the neighbor list of the start node,
and not as a standalone entry in the adjacency list itself. This behavior has been fixed