All addEdge member functions now behave the same
Previously only the addEdge that takes a GraphClasses::Edge struct as a parameter inserted
one direction of edge, while the other two overloads checked if the graph is undirected and inserted both directions.
This behavior was confusing and could possibly lead to bugs when implementing featurs in the future.
Now all three addEdge overloads insert only one direction of an undirected edge and it is up to the user to call
the same function again (with switched places of parameters) to insert the other direction